backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend for search-take-home task"
requires-python = ">=3.11, <3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.121.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain[openai]>=1.2.10",
    "numpy>=2.3.4",
    "uvicorn>=0.38.0",
]
